+++ release-notes.md
@@ -4,6 +4,12 @@ for [Chrome extension to for Bible.com](README.md) by [@Matei Nicolae](https://n
 
 💚 **Love this extension?** Share [feedback](https://chromewebstore.google.com/detail/project-verses-from-bible/fklnkmnlobkpoiifnbnemdpamheoanpj) and help us make it even better + helps others to find this resource.
 
+## 2.12.0 (comming soon)
+
+- [x] Updated marked library from v15.0.6 to v17.0.1 (latest Markdown parser)
+- [x] Added DOMPurify v3.3.0 for HTML sanitization (security enhancement)
+- [x] Allow external extensions to send updateText messages to projector tab (with text and markdown flag)
+
 ## 2.11.0 (2025-10-07)
 
 - [x] Allow different design (slide master) on each window - from settings button select Window 1 or Window 2 then "Slide:" menu to select design for that particular window (eg. Window 1 = Projector, Window 2 = TV)